+++ projects/ngx-mask-lib/src/test/export-as.spec.ts
@@ -1,9 +1,9 @@
 import { ComponentFixture, TestBed } from '@angular/core/testing';
-import { NgxMaskModule } from '../lib/ngx-mask.module';
 import { ReactiveFormsModule } from '@angular/forms';
 
 import { Component, ViewChild } from '@angular/core';
-import { MaskDirective } from 'ngx-mask';
+import { provideNgxMask } from '../lib/ngx-mask.providers';
+import { NgxMaskDirective } from '../lib/ngx-mask.directive';
 
 @Component({
     selector: 'mask-test-mask',
@@ -11,10 +11,10 @@ import { MaskDirective } from 'ngx-mask';
 })
 export class TestMaskComponent {
     @ViewChild('refMask')
-    public refMask!: MaskDirective;
+    public refMask!: NgxMaskDirective;
 
     @ViewChild('refNgxMask')
-    public refNgxMask!: MaskDirective;
+    public refNgxMask!: NgxMaskDirective;
 }
 
 describe('Directive: Mask export As', () => {
@@ -24,7 +24,8 @@ describe('Directive: Mask export As', () => {
     beforeEach(() => {
         TestBed.configureTestingModule({
             declarations: [TestMaskComponent],
-            imports: [ReactiveFormsModule, NgxMaskModule.forRoot()],
+            imports: [ReactiveFormsModule, NgxMaskDirective],
+            providers: [provideNgxMask()],
         });
         fixture = TestBed.createComponent(TestMaskComponent);
         component = fixture.componentInstance;
@@ -32,9 +33,9 @@ describe('Directive: Mask export As', () => {
     });
 
     it('should export directive instance with exportAs api by name equal mask', () => {
-        expect(component.refMask).toBeInstanceOf(MaskDirective);
+        expect(component.refMask).toBeInstanceOf(NgxMaskDirective);
     });
     it('should export directive instance with exportAs api by name equal ngxMask', () => {
-        expect(component.refNgxMask).toBeInstanceOf(MaskDirective);
+        expect(component.refNgxMask).toBeInstanceOf(NgxMaskDirective);
     });
 });