+++ src/app/bugs/shadow-dom/shadow-dom.component.ts
@@ -1,8 +1,11 @@
 import { Component, ViewEncapsulation } from '@angular/core';
+import { NgxMaskDirective } from 'ngx-mask';
 
 @Component({
     selector: 'mask-shadow-dom',
     templateUrl: './shadow-dom.component.html',
+    standalone: true,
+    imports: [NgxMaskDirective],
     encapsulation: ViewEncapsulation.ShadowDom,
 })
 export class ShadowDomComponent {}